Updated auto.py with for-loops instead of Numpy array ops. - 200K nodes / sec!!!
In this update I managed to fix the memory leak caused by Numba's generators - largely documented in the following links:

- https://github.com/numba/numba/issues/5427
- https://github.com/numba/numba/issues/5350
- https://github.com/numba/numba/issues/6993
- https://github.com/numba/numba/issues/5350

Now I simply pre-allocate an array with the maximum number of moves possible for a player in the current state.

What's next? - I am going to try and pre-allocate all resources needed to execute the search - except for a few.